rm circular type dep
