+++ package.json
@@ -9,7 +9,8 @@
     "preinstall": "npx only-allow pnpm",
     "wipe": "npkill",
     "build": "pnpm -r --parallel --filter !playground run build",
-    "build:ci": "pnpm -r --filter !playground run build",
+    "build:types": "pnpm -r --filter !playground exec -- tsup --dts-only",
+    "build:ci": "pnpm build:types && pnpm build",
     "dev": "pnpm -r --parallel --filter !playground run dev",
     "lint": "prettier --check --plugin-search-dir=. .",
     "lint:fix": "pnpm lint:fix:prettier",
