+++ packages/flatbread/package.json
@@ -32,7 +32,7 @@
     "dist",
     "*.d.ts"
   ],
-  "dependencies": {
+  "devDependencies": {
     "@flatbread/config": "workspace:^1.0.0-alpha.4",
     "@flatbread/core": "workspace:^1.0.0-alpha.11",
     "@flatbread/source-filesystem": "workspace:^1.0.0-alpha.4",
@@ -50,9 +50,7 @@
     "kleur": "^4.1.5",
     "remark-github": "^11.2.3",
     "sade": "^1.8.1",
-    "serialize-javascript": "^6.0.0"
-  },
-  "devDependencies": {
+    "serialize-javascript": "^6.0.0",
     "@types/express": "^4.17.13",
     "@types/gradient-string": "^1.1.2",
     "@types/node": "^16.11.42",
