+++ packages/transformer-yaml/package.json
@@ -33,16 +33,14 @@
   "engines": {
     "node": "^14.13.1 || >=16.0.0"
   },
-  "dependencies": {
+  "devDependencies": {
     "@sindresorhus/slugify": "^2.1.0",
     "graphql": "16.0.1",
     "gray-matter": "^4.0.3",
     "js-yaml": "^4.1.0",
     "lodash-es": "^4.17.21",
     "to-vfile": "^7.2.3",
-    "unified": "^10.1.2"
-  },
-  "devDependencies": {
+    "unified": "^10.1.2",
     "@flatbread/core": "workspace:^1.0.0-alpha.11",
     "@types/js-yaml": "^4.0.5",
     "@types/node": "^16.11.42",
